Create Finite Difference Methods - Data Visualization.py
The finite-difference method can easily be extended to higher dimensions by using
the finite-difference formula along each discretized coordinate. For a two-dimensional
problem, we have a two-dimensional array u for the unknown interior function values,
and when using the finite differential formula, we obtain a system of coupled equations
for the elements in u. To write these equations on the standard matrix-vector form, we
can rearrange the u array into a vector and assemble the corresponding matrix A from
the finite-difference equations.